+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## v1.0.3
+
+- Dynamically add class name base on tag-name:
+  - _**INPUT**_: `field__input`
+  - _**TEXTAREA**_: `field__textarea`
+  - _**P-DROPDOWN**_': `field__dropdown`
+
 ## v1.0.2
-* Project Creation
+
+- Project Creation